refactor: 회원 도메인 레이어 구조 변경 (#64)
