--- src/main/kotlin/kr/wooco/woocobe/user/infrastructure/storage/UserJpaRepository.kt
@@ -1,5 +1,0 @@
-package kr.wooco.woocobe.user.infrastructure.storage
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserJpaRepository : JpaRepository<UserEntity, Long>
