+++ src/main/kotlin/kr/wooco/woocobe/user/infrastructure/storage/entity/UserEntity.kt
@@ -1,4 +1,4 @@
-package kr.wooco.woocobe.user.infrastructure.storage
+package kr.wooco.woocobe.user.infrastructure.storage.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
