+++ src/main/kotlin/kr/wooco/woocobe/user/ui/web/controller/request/UpdateUserProfileRequest.kt
@@ -1,4 +1,4 @@
-package kr.wooco.woocobe.user.ui.web.dto.request
+package kr.wooco.woocobe.user.ui.web.controller.request
 
 data class UpdateUserProfileRequest(
     val name: String,
